Consolidate ContentClient into a single common class
