+++ apps/checker/app/AppComponents.scala
@@ -30,7 +30,7 @@ import play.filters.HttpFiltersComponents
 import play.filters.cors.CORSComponents
 import router.Routes
 import services._
-import com.gu.typerighter.lib.Loggable
+import com.gu.typerighter.lib.{ContentClient, Loggable}
 import com.gu.typerighter.rules.BucketRuleResource
 import matchers.LanguageToolFactory
 import utils.CloudWatchClient
