+++ apps/checker/app/controllers/CapiProxyController.scala
@@ -4,8 +4,7 @@ import com.gu.contentapi.json.CirceEncoders._
 import io.circe.syntax._
 import play.api.mvc._
 import com.gu.typerighter.controllers.PandaAuthController
-import com.gu.typerighter.lib.CommonConfig
-import services.ContentClient
+import com.gu.typerighter.lib.{CommonConfig, ContentClient}
 
 import scala.concurrent.ExecutionContext
